Helper Correction

+++ app/Http/Helpers.php
@@ -28,7 +28,7 @@ function calcValuePercentPos($price, $percent)
 
 function calcPercent($price, $percent)
 {
-    $percent = $price * $percent / 100;
+    $percent = $price * $percent / (int)100;
     $result = convertAndRoundDecimal($percent, 2);
     if($result == '0'){
         $result = '0.00';
